redesign and fix on login page
