+++ Frontend/frontend/src/app/store.js
@@ -1,6 +1,6 @@
 import { configureStore } from "@reduxjs/toolkit";
 import { apiSlice } from './api/apiSlice';
-import authReducer from '../features/auth/authSlice';
+import authReducer from '../components/authSlice';
 
 export const store = configureStore({
     reducer: {
