+++ Frontend/frontend/src/components/Login.js
@@ -1,30 +1,104 @@
-import { useNavigate } from 'react-router-dom'
-import React, { useState } from "react";
+import { useRef, useState, useEffect } from 'react'
+import { useNavigate, Link } from 'react-router-dom'
+
+import { useDispatch } from 'react-redux'
+import { setCredentials } from './authSlice'
+import { useLoginMutation } from './authApiSlice'
+import "./Menu.css";
+
 
 const Login = () => {
-    const navigate = useNavigate();
+    const userRef = useRef()
+    const errRef = useRef()
+    const [username, setUsername] = useState('')
+    const [password, setPassword] = useState('')
+    const [errMsg, setErrMsg] = useState('')
 
+    const navigate = useNavigate()
+    const dispatch = useDispatch()
 
-    const [email, setEmail] = useState('');
-    const [pass, setPass] = useState('');
+    const [login, { isLoading }] = useLoginMutation()
 
+    useEffect(() => {
+        userRef.current.focus()
+    }, [])
 
-    const handleSubmit = (e) => {
+    useEffect(() => {
+        setErrMsg('');
+    }, [username, password])
+
+    const handleSubmit = async (e) => {
         e.preventDefault();
-        console.log(email);
-        navigate("/menu");
+        try {
+            const { accessToken } = await login({ username, password }).unwrap()
+            dispatch(setCredentials({ accessToken }))
+            setUsername('')
+            setPassword('')
+            navigate('/menu')
+        } catch (err) {
+            if (!err.status) {
+                setErrMsg('No Server Response');
+            } else if (err.status === 400) {
+                setErrMsg('Missing Username or Password');
+            } else if (err.status === 401) {
+                setErrMsg('Unauthorized');
+            } else {
+                setErrMsg(err.data?.message);
+            }
+            errRef.current.focus();
+        }
     }
+
+    const handleUserInput = (e) => setUsername(e.target.value)
+    const handlePwdInput = (e) => setPassword(e.target.value)
+
+    const errClass = errMsg ? "errmsg" : "offscreen"
+
+    if (isLoading) return <p>Loading...</p>
+
     const content = (
-        <>
-                <h1>Login</h1>
-                <form className="login-form" onSubmit={handleSubmit}>
-                <label htmlFor="email">email</label>
-                <input value={email} type="email" onChange={(e) => setEmail(e.target.value)} placeholder="youremail@gmail.com" id="email" name="email" />
-                <label htmlFor="password">password</label>
-                <input value={pass} type="password" onChange={(e) => setPass(e.target.value)} placeholder="********" id="password" name="password" />
-                <button type="submit">Log In</button>
+        <section className='public'>
+        <header>
+            <h1>Login</h1>
+        </header>
+        <main className='login'>
+            <p ref={errRef} className={errClass} aria-live="assertive">{errMsg}</p>
+            <form className="login-form" onSubmit={handleSubmit}>
+                <div className="menuButtonContainer">
+                    <label htmlFor="username">Username: </label>
+                    <input 
+                        className='form_input'
+                        type='text'
+                        id='username'
+                        ref={userRef}
+                        value={username}
+                        onChange={handleUserInput}
+                        autoComplete="off"
+                        required
+                        placeholder="username
+                        "
+                        name="username" 
+                    />
+                    <br/>
+                    <label htmlFor="password">Password: </label>
+                    <input 
+                        className='form_input'
+                        type='password'
+                        id='password'
+                        onChange = {handlePwdInput}
+                        value={password}
+                        required
+                        placeholder="********"
+                        name="password"
+                        />
+                    <br/>
+                    <button className="menuButton">
+                        Sign In
+                    </button>
+                </div>
             </form>
-            </>
+        </main>
+    </section>
         );
         return content;
 };
