+++ Frontend/frontend/src/components/authApiSlice.js
@@ -1,4 +1,4 @@
-import { apiSlice } from "../../app/api/apiSlice"
+import { apiSlice } from "../app/api/apiSlice"
 import { logOut } from "./authSlice"
 
 export const authApiSlice = apiSlice.injectEndpoints({
